rename todolist to todo
